requirements.txt
python-telegram-bot==21.6
python-dotenv>=1.0.0

# AI / LLM
langchain>=1.2.0
langchain-ollama>=1.0.0
langchain-community>=0.4.0

# Google Calendar
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Therapist web frontend
fastapi>=0.110.0
uvicorn>=0.29.0
jinja2>=3.1.0
python-multipart>=0.0.9

# Future — Anthropic (uncomment when switching from Ollama)
# anthropic>=0.40.0
